+++ include/ax.php
@@ -1,0 +1,11 @@
+<?php addHeading("Advanced X", "about"); ?>
+
+<section>
+	<p><em>Advanced X</em> (or just <em>AX</em>) is a C library for developing apps for the <em>AGB</em> line of hardware.</p>
+</section>
+
+<?php addHeading("Status", "status"); ?>
+
+<section>
+	<p>The development of <em>AX</em> is currently on hold to prioritise <a href="?p=luma"><em>Luma</em></a> and <a href="?p=eas"><em>eAS</em></a>.</p>
+</section>