+++ include/backspace.php
@@ -1,0 +1,3 @@
+<section class="fullWidth">
+	<p class="note">coming soon</p>
+</section>