+++ include/dux.php
@@ -1,0 +1,11 @@
+<?php addHeading("Dux", "about"); ?>
+
+<section>
+	<p><em>Dux</em> is a cross-platform widgeting library for developing GUI applications.</p>
+</section>
+
+<?php addHeading("Status", "status"); ?>
+
+<section>
+	<p>Dux is currently on standby and therefore unmaintained.</p>
+</section>