+++ include/luma.php
@@ -1,0 +1,11 @@
+<?php addHeading("Luma", "about"); ?>
+
+<section>
+	<p><em>Luma</em> is an emulator for the <em>AGB</em> line of game consoles by Nintendo Co.</p>
+</section>
+
+<?php addHeading("Status", "status"); ?>
+
+<section>
+	<p>The development of <em>Luma</em> is currently on hold in favour of <a href="?p=pollex"><em>Pollex</em></a>, which is to be used as a backend.</p>
+</section>