+++ include/pollex.php
@@ -1,0 +1,15 @@
+<?php addHeading("Pollex", "about"); ?>
+
+<section>
+	<p><em>Pollex</em> is a Rust crate for manipulating Arm ISA instructions. Its goal is to be used as a backend for Arm emulators, assemblers etc.</p>
+	<br>
+	<p>Currently, only the legacy 32-bit architectures are supported (if barely), but complete support is planned in the future.</p>
+	<br>
+	<p>See more at <a href="https://crates.io/crates/pollex/"><code>crates.io</code></a>.</p>
+</section>
+
+<?php addHeading("Docs", "docs"); ?>
+
+<section>
+	<p>As per usual, documentation for <em>Pollex</em> can be found on <a href="https://docs.rs/pollex/latest/pollex/"><code>docs.rs</code></a>.</p>
+</section>