+++ include/u8c.php
@@ -1,0 +1,25 @@
+<?php addHeading("u8c", "about"); ?>
+
+<section>
+	<p><em>u8c</em> is a library for handling Unicode sequences in C.</p>
+</section>
+
+<?php addHeading("Features", "features"); ?>
+
+<section>
+	<p><em>u8c</em> supports the following facilities:</p>
+	<br>
+	<ul>
+		<li>
+			<p>Encoding and decoding <em>UTF-8</em></p>
+		</li>
+		<li>
+			<p>Encoding and decoding <em>UTF-16</em></p>
+		</li>
+		<li>
+			<p>Manipulation of <em>UTF-32</em> strings.</p>
+		</li>
+	</ul>
+	<br>
+	<p>Note that <em>UTF-32</em> is used as an intermediate format between conversions.</p>
+</section>