+++ install.sh
@@ -18,9 +18,9 @@ mkdir -p "${DESTINATION}"
 
 cp "apple-touch-icon.png" "${DESTINATION}"
 cp "favicon.ico" "${DESTINATION}"
+cp "index.php" "${DESTINATION}"
 
 cp -r "font" "${DESTINATION}"
-cp -r "html" "${DESTINATION}"
 cp -r "image" "${DESTINATION}"
 cp -r "include" "${DESTINATION}"
 cp -r "svg" "${DESTINATION}"